refactor: Mark notification as read feature
- integrate `Notification` model to `handleNotifications.js`.
- move the marking notification as read logic to `handleNotificationEvents` handler for `mark as read` socket event in `handleNotifications.js`.
- remove the redundant `updateNotification` controller function in `notificationController.js`.
- remove the redundant integration of `updateNotification` an the matching API route endpoint in `notificationRoutes.js`.
